src/main/java/CzescII/UserRegistration.java
package CzescII;

import java.util.ArrayList;
import java.util.List;

public class UserRegistration {

    private List<User> userList = new ArrayList<> ();



    public boolean registerNewUser(String name, String email, String password, String passwordConfirmed, UserType userType) {
        if(password.equals (passwordConfirmed)&& isMailValid (email)){
        User user = new User (name,email,userType);
        userList.add (user);
        return true;
        }
        return false;
    }

    public List<User> getUserList() {
        return userList;
    }

    public void setUserList(List<User> userList) {
        this.userList = userList;
    }

    public boolean isMailValid(String email) {
        String emailRegex = "(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|\"(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21\\x23-\\x5b\\x5d-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])*\")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\\[(?:(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9]))\\.){3}(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9])|[a-z0-9-]*[a-z0-9]:(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21-\\x5a\\x53-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])+)\\])";
          return email.matches (emailRegex);
    }

    public boolean isPasswordValid(String password, String passwordConfirmed) {
        if (password.equals (passwordConfirmed)){
            return true;
        }
        return false;
    }
}
